Add orbit camera and input system
Adds a simple orbit camera based on spherical coordinates. The camera
supports:

- panning in the screen plane (rmb + movement)
- panning in the XY plane     (shift + rmb + movement)
- rotation around origin      (lmb + movement)
- step zoom                   (wheel)
- smooth zoom                 (lmb + rmb + movement)
- origin reset                (A)

The input system is factored out to a separate struct:
InputManager. Each frame, the manager needs to be reset via
`start_frame()` and filled with events it needs to process
via `process_event()`. Afterwards, a pointer to input state can be
queried.