Add matcap shading
This modifies the viewport renderer to shade the geomeotry with the material
capture (matcap) technique. Matcap entails using the .xy components of
a viewspace projected normal to sample a texture containing
pre-renderered lighting information. Note that the texture included in
this PR is not very good and we should create a better one later.

**Some changes:**

There is a new `uv_cube()` primitive geometry. All primitives
live in RH coordinate space with XY being the ground and Z growing upwards.

Matcap means we now always need the geometry normals present (and will
have to compute them if they are not). `Geometry` struct is created to
encapsulate what the renderer needs. Its `indices` field is optional.

The renderer currently decodes the matcap texture during
initialization, and the texture is baked to the program binary (might
want to change that to runtime loading later).

Since we need to also use the view matrix on the shaders to transform
object normals to view space, camera now needs to export its matrices
separately.