Implement MSAA antialiasing for viewport renderer
This commit updates wgpu-rs and gfx-hal to 0.3 (needed because MSAA
didn't work on metal with 0.2) and implements MSAA for viewport renderer.

With gfx-hal 0.3 MSAA now works, but 3 different stencil-state vulkan
validation messages popped up:

- stencil reference is not set (this commit "fixes" that by setting
  it)
- stencil read mask is not set
- stencil write mask is not set

The last two don't look like they can be set with the current API. No
real problems that look related to these validation messages manifest
while the program is running on any of our backends, though.

Multi-sampling itself is implemented by the renderer allocating a MSAA
framebuffer texture ans using it as the render target while using the
original render target as the MSAA resolve target. This is only done
if multi-sampling is enabled, which is an option in viewport renderer.

This commit leaves the application in a temporary state - MSAA should
happen only once and for both the GUI and the window renderers, but
the GUI renderer is not present at the time of writing. We'll have to
refactor later.