Add zoom-to-fit (bounding sphere)
To add the functionality, some inspectable geometry representation
needed to be created (previous renderer geometry has all fields module
private). I took this as an opportunity to start sketching what the
"editor geometry" structure would look like.

Former `primitives` module has been moved to the `geometry` module,
and the primitive constructors now create the editor geometry. A
conversion exists from editor geometry to renderer geometry.

Zoom-to-fit itself is implemented by finding a bounding sphere of the
scene and using it to change zoom and origin (but not the angle) on
the camera.