Implement edge calculation, edge analytics and edge-based mesh analytics
IMPLEMENT EDGE CALCUALTION
Edges are calculated from Faces as linear connections of their three Vertices. 
OrientedEdge preserves face winding, UnorientedEdges don't (lower vertex index first), however two (or more) adjacent faces produce two (or more) UnorientedEdges, too.
Wireframe edges (inner edges as a single edge) are partly implemented as SimilarEdges and they are also calculated ad-hoc as a HashSet of UnorientedEdges.

IMPLEMENT EDGE-BASED MESH ANALYTICS
EdgeCountMap under a key of UnorientedEdge contains original OrientedEdges. This is useful for edge valency (number of connections to faces) calculations: valency=1 => edge is on the border of the mesh, valency=2 => edge is inside mesh (is manifold), valency>2 => edge is shared by more than 2 faces (is non-manifold). 
An acceptable mesh consists of valency=1 and valency=2 edges and a perfectly valid watertight mesh only consists of valency=2 edges (has no borders or non-manifold details).
Mesh is orientable (has a consistent inner and outer side) when the shared edge of two faces is actually made of two differently oriented edges.
Mesh genus is a number of holes in the topology of a valid watertight mesh. A sphere (or a cube) is genus 0, a donut (or a mug) is genus 1, etc.