Implement interpreter with variable cache and invalidation (#35)
This is the first iteration of our Rooty-Lang interpreter. What's
missing:

- The ability to pause and resume the interpreter
- Iterative funcs a.k.a. Generators

Be aware that there is the barest amount of program
verification implemented, and that is mostly runtime. Compile time
verification is yet to come. Invalid programs may crash the
interpreter.

The interpreter is integrated into the editor with `InterpreterServer`,
which works very similarly to what the Importer worker does.