Laplacian smoothing
This commit implements Laplacian smoothing routine and creates an interpreter operation utilizing it.
Laplacian smoothing (relaxation) is an iterative mesh transformation resulting in a smoother mesh without adding new vertices or faces. Each vertex is moved to an average position of its neighbors. The routine is usually applied in several (3-10) consecutive iterations, too many iterations, however, cause the model to "disappear" into a small round amorphous goo.
The tests of the Laplacian smoothing use data snapshots of the same algorithm implemented in Grasshopper. The tests also prove the vertex-to-vertex topology computation works as expected.