Find mesh border edge loops
Find mesh border edge loops, which are circular sequences of edges (closed polylines) around the mesh border. Border edges are contained exclusively in a single face. The feature will be later needed for parametrization of meshes before morphing, for splitting into patches and for physical simulations we might need in some morphing algorithms.

A mesh can have:

no border edge loop, when it's watertight
single border edge loop, when it's a patch or a mesh with a single hole (topologically also a patch)
more border edge loops, when it has more holes or consists of more mesh patches.
A border edge loop always consists of 3 or more border edges (for mesh patches made of a single triangular face or for meshes with a single triangle missing).

In an edge case when more border edge loops share a single vertex (checkerboard scenario), there is no resolution strategy, so the result is left unpredictable, yet correct. In the future, the function can detect the case and throw a warning.

Also, in this commit most of the valency-related types and functions are renamed more suitably.