Remove triangle_faces_iter and triangle_faces_count from Geometry (#46)
It seems this filtering view is virtually never what we want to do,
when we ever have nontriangular faces. It was allowing us to not care
about other kind of faces, potentially leading to incorrect behavior
in the future.

This also adds two new constructors:

- from_faces_with_vertices_and_normals
- from_faces_with_vertices_and_normals_remove_orphans

These don't take triangle faces, but the generic enum `Face` instead,
to make for easier geometry construction after `triangle_faces_iter`
was removed.

Additionally, this commit generifies all `Geometry` constructors to
take `IntoIterator`s instead of collections to save some allocations
from happening, as the various constructors chain on each other.

There is currently no normal-less alternative to the two functions
introduced here and hopefully we won't need it, as we don't want to
describe all kinds of faces once without normals and once with them.