Operation: Laplacian smoothing
Implement Laplacian smoothing routine and create an interpreter operation utilizing it.

Laplacian smoothing (relaxation) is an iterative mesh transformation resulting in a smoother mesh without adding new vertices or faces. Each vertex is moved to an average position of its neighbors. The routine is usually applied in several (0-255) consecutive iterations. Too many iterations, however, cause the model to "disappear" into a small round amorphous goo or into a singularity.
The tests of the Laplacian smoothing use data snapshots of the same algorithm implemented in Grasshopper. The tests also prove the vertex-to-vertex topology computation works as expected.