Compare visually similar geometries
Check if two mesh geometries are identical

Two mesh geometries are identical when the position of each vertex in one mesh geometry matches a position of exactly one vertex in the other mesh geometry, when the direction of each normal in one mesh geometry matches a direction of exactly one normal in the other mesh geometry and each face in one mesh geometry refers vertices with the same position and normals with the same direction, both in the same circular order, as exactly one face in the other mesh geometry. The indices (order in which they are stored) of vertices, normals and faces can differ but as long as the previous conditions are met, the mesh geometries are identical.