Stop laplacian smoothing with anchors when the geometry becomes stable
The Laplacian smoothing with anchors can be used to paramterize a mesh patch into a flat disc. Once the edge loop is transformed into the circumference of the target disc, the iterative algorithm of Laplacian smoothing is applied until the mesh reaches a stable spatial configuration.

This function will be later used in more complex operations, therefore it's not exposed as a distinct operation.