Add ability to draw mesh edges and add more drawing modes
This adds 4 drawing modes and the ability to switch between them:

- Shaded:          What we had until now
- Edges:           Draws just the edges of the mesh
- ShadedEdges:     Draws both the shaded view and the visible edges
- ShadedEdgesXray: Draws both the shaded view and all the edges, even
                   occluded

The edge drawing is done with a shader trick, which seemed like the
simplest solution initially (required the least re-architecting), but
it might not be enough for our later needs. See `viewport.frag` for
more.

The scene renderer is refactored internally to support
(very trivial and mostly incorrect) rendering of transparent objects:
`Edges` and `ShadedEdgesXray` use the new transparent render pipeline.

Both rendering of faces and edges is done by the same shader, with
flags used to enable rendering of one or the other, or both (for
`ShadedEdges`). `ShadedEdgesXray` is rendered in two passes - first
`Shaded`, then `Edges`.