Distinct between identical, similar and visually similar mesh geometries
This is a more rigorous comparison of two mesh geometries.

Let's consider two mesh geometries IDENTICAL when the data structure looks the same and they look the same in the memory. A simple Eq function should suffice for this comparison.

Let's consider two mesh geometries SIMILAR when they look exaclty the same, have the same features and behave the same but don't necesarilly look the same in memory. Such mesh geometries share the same count of vertices, normals and faces but they can be sotred in different orders. Faces may reference the vertices in different order but identical winding (0->1->2 = 1->2->0 = 2->0->1).

Let's consider two mesh geometries VISUALLY SIMILAR when they look exaclty the same, but neither necesarilly have the same features, nor behave the same, nor look the same in memory. This is usually the case when faces don't reuse (share) the same vertex more times but rather reference an individual vertex with the same coordinates for each of the faces. This can be the case for all the faces/vertices or just some. Such meshes look exactly the same but they are treated differently by this software's functions and have different features - usually at least one of them is not watertight. Not reusing normals doesn't result in different behavior or features but for the sake of this distinction two meshes with different counts of normals are also considered VISUALLY SIMILAR but not SIMILAR.

This distinction is important when using some of our mesh tools. separate_isolated_meshes applied to a mesh geometry, which consists of two or more isolated chunks sharing the same normals, yields two geometires with some of the normals copied to both geometries. Using join_meshes on these separated chings later on doesn't yield the original mesh but rather one with the originally shared normals now copied as individual elemnts referenced by only one chunk. Such joined mesh is VISUALLY SIMILAR to the original one but is not SIMILAR. The same would apply to vertices when a mesh geometry is being chopped into more pieces and then joined again or to unwelded and welded versions of the same geometry.