Weld mesh geometry vertices and normals
Weld similar (close enough) vertices into one and reuse such vertices in connected faces

Weld is used to actually connect faces, which are connected merely visually (often resulting in a watertight mesh) or to reduce number of vertices in case when vertices are multiplied because vertices at the same coordinates are not referenced by more faces but rather each face references its own copy of the vertex.

Weld is one of the auto-fixes leading to a simplified, watertight or true-to-its-genus mesh geometries.