Add basic name resolution (#42)
This allows us to remove a few runtime errors, and make them
"compile-time" errors instead. When we decide to do static
typechecking in the future, this can act as the guideline for adding
it.

As far as the frontend/UI is concerned, name resolution errors should
ideally never happen. Maybe the sensible strategy is to panic on
these, but otherwise make sure correct programs are constructed.