Synchronize and revert mesh face winding
SYNCHRONIZE MESH FACE WINDING

Make sure all the faces are oriented the same way - have the same winding (vertex order).

This function crawls the mesh geometry and flips all the faces, which are not facing the same way as the previous faces in the process, starting with the first face in the list. As a result, the entre mesh can end up facing inwards (be entirely reverted). At the moment we have no tools to detect such a case automatically, so we need to rely on the user to check it and potentially revert winding of the entire mesh.

The algorithm relies on the fact that in a proper non-manifold mesh, each oriented edge has exactly one (for watertight mesh geometry) or none (for mesh patches) counter-edge oriented the opposite direction. It crawls the geometry and if a face neighboring the current one doesn't have the proper winding, it's being reverted and only then triggers checking its own neighbors.

This method, doesn't flip the normals associated with the face vertices, as there is no unambiguous way to do so automatically.

REVERT MESH FACE WINDING

Reverts vertex and normal winding of all faces in the mesh geometry and returns a reverted mesh geometry