Pull point to mesh
PULL POINT TO MESH
Pulls arbitrary point to the closest point of a mesh geometry.
Cast a ray from the point perpendicular to each mesh face and if there is an intersection, measure the distance. Also measure a distance to the closest point on each mesh geometry edge. Pick the closest point of them all as the pulled point.

PULL POINT TO LINE
Pulls arbitrary point to an arbitrary line.
Returns both, a clamped point on the line adn an unclamped points on an endless ray.
https://stackoverflow.com/questions/3120357/get-closest-point-to-a-line

RAY INTERSECTS TRIANGLE
The Möller–Trumbore ray-triangle intersection algorithm is a fast method for calculating the intersection of a ray and a triangle in three dimensions without needing precomputation of the plane equation of the plane containing the triangle.
https://en.wikipedia.org/wiki/Möller–Trumbore_intersection_algorithm
https://www.scratchapixel.com/lessons/3d-basic-rendering/ray-tracing-rendering-a-triangle/moller-trumbore-ray-triangle-intersection

CALCULATE TRIANGLE FACE NORMAL
Calculate a normalized vector perpendicular to the face pointing outwards following the right-hand rule.