Add ability to make snapshot tests
Snapshot tests allow to detect changes in output of complex data, in
our case geometry. When an output of a tested function changes (or the
functionality is new and there is no snapshot to compare against yet),
the data has to be manually reviewed and approved. For us, that will
most likely mean comparing with other 3d software, or at least
visually checking in the HS editor.

Snapshot testing is provided by the `insta` library. Snapshot review
can also be done manually, but using the cargo plugin `cargo-insta`
simplifies snapshot review - `cargo insta review` interactively lists
changed snapshots and asks the reviewer to accept or reject them.

This removes a quite verbose laplacian smoothing test and replaces it
with snapshots. I also noticed one spot, where topology is passed by
value and unnecessarily cloned all over the place, so I fixed that.