Remove cube_sharp_same_len
Initially there was a fixme to check whether the gpu geometry
conversion was correct and keep this until we verify it, but
realistically we won't have time to do that any time soon and so far
there have been no problems with the conversion.

I also renamed cube_smooth_same_len to just cube_smooth. We now have
some experience of working with variable length data in geometry, and
it doesn't seem like it's going to pose any problems, so from now on
we assume everything is (potentially) variable length.