Add ability to return runtime errors from funcs (#69)
This changes the return type of funcs to `Result<Value, FuncError>`,
where `FuncError` wraps a dynamic error type. When a func returns an
error, the whole interpreter session returns the same error under the
newly created `RuntimeError::Func` variant.

Due to the dynamic nature of `FuncError`, `PartialEq` for it is
implemented with strict pointer equality. (If we wanted to implement
this by doing a complete equality check, the fact that trait objects
don't implement `PartialEq` would make it rather challenging).