Pull vertex (#71)
The pull functions are inspired by their equivalents in Grasshopper - they finds the closest point on the respective geometry to current point. This is a non-trivial logic, because the behavior is different for various target geometries. 

Pull to point returns the target point.
Pull to line returns a point on the line, which is a perpendicular projection of the current point onto the line or one of the line ends in case the perpendicular projection doesn't fall into the domain of the line.
Pull to plane returns a perpendicular projection of the current point onto the plane.
Pull point to face returns a perpendicular projection of the current point onto a triangular face or a perpendicular projection of the current point onto a face edge (in case the perpendicular projection onto the triangle plane doesn't fall into the are of the triangle) or one of the triangle corners (in case the perpendicular projection of the current point doesn't fall into a domain of any of the triangle face edges).

The pull functions return the pulled (closest) point and the distance between it and the original point.

The functionality is vast and brings along many supplemental functions as well as data types. The most important is a Plane, which is a custom orthogonal coordinate system with arbitrary origin and rotation (normal) but original scale and no spatial distortion.