Mesh groups (#80)
This adds mesh groups... or `Value::GeometryArray`. In the interpreter, they are just another value. In the UI and viewport, we need to know about them specially to display them in the combo boxes and the viewport. Current user-visible behavior in viewport is identical to other geometries - there is no way to tell those 2 apart (probably will improve in at some point). In the UI combo boxes, each combo box only receives suggestions for the type it expects - one can not accidentally pass a `GeometryArray` where `Geometry` is expected and vice versa.

Two existing funcs were changed to return `GeometryArray`: `FuncImportObj`, and `FuncSeparateIsolatedMeshes`. To get a geometry from an array, use the newly added `FuncExtract`.

There seems to be functionality duplication growing between the session and the interpreter (mostly because they can only communicate asynchronously and the session doesn't like waiting for answers). It is not too terrible yet, but it would be a good target for some future cleanup effort. There is a new fixme for that.