Fix and optimize weld function (#82)
There was an error in the weld function. The newly created faces could happen to consist of two or three identical vertex indices, which would be invalid. Therefore there was a filter to remove such faces but it came AFTER the attempt to create the faces. The face constructor panicked (without any console output and without crashing the program for some reason) and the weld has never finished.

There also was an unnecessary HashMap, so it's replaced by a vector now. The initial value for the indices is a value, that will cause the rest of the code panic if any of the vertices remained untreated.