Fix weld generating zero vector normals in some cases
The weld function calculates normals as an average of all normals associated to a respective vertex (a vertex can be referenced by multiple faces, each of them may couple it with a different normal). The original weld function deduplicated associated normals, which was not correct. 
Some optimizations were made too and the tests were adjusted to the current algorithm.