Add interpreter log messages
Each func run in the interpreter now has the option to log
messages. These messages are returned in the interpreter's response
(`InterpretOutcome`) regardless of whether the interpretting ended
successfully. The UI is updated with autoscrolling consoles that
display the log messages. On the UI, the log messages are always
appended to the console output and the consoles are only cleared when
the operation is removed from the pipeline.

The interpreter appends one log message with the func's run duration
to the func's log after it returns. For cached (and not invalidated)
func results, no new logs are produced, but the interpreter appends a
log message saying that the result was taken from cache. If the func
ends with an error, the interpreter also appends a log message (with
the `Error` level).

Minor refactoring happened in both the interpreter and the UI:

- Interpreter no longer keeps track of the program counter in between
calls to `Interpreter::interpret{_up_until}()` and instead the PC
always begins at 0. This was the simplest solution for producing log
messages for operations we would otherwise skip by not resetting the
PC back to zero after invalidation. This will have a very minor
performance impact (the `env` hash map will be accessed more times)
and a FIXME was left in the code to address the problem in a more
systematic manner later.

- Since there was no way around adding a `RefCell` around UI state for
the consoles, a global string buffer wrapped in a `RefCell` for imgui
components to use is also added to the `Ui` as an
optimization. Rammifications should be minimal, we don't expect to
ever want to `borrow_mut` these `RefCell`s in a reentrant manner. UI
performance should improve slightly.