Add voxel-based boolean operations on mesh geometry
All operations first convert the geometry into voxel clouds , then perform the logical operations and eventually convert the voxel clouds into meshes. 

Intersection: boolean AND - keeps those voxels, that are set to true in both voxel cloud operands
Union: boolean OR - keeps those voxels, that are set to true in either voxel cloud operands
Difference: keeps those voxels, that are set to true in the first voxel cloud operand but not set to true in the second

This commit implements the Intersection and Union operations on two voxel clouds, one of which is being mutated and on a group (slice of voxel clouds), when a new voxel cloud is being created.