Add ability to screenshot the viewport (#116)
* Add ability to screenshot the viewport

The functionality is triggered by either the P key, or by clicking on
the "screenshot" button in the "viewport" window. That brings up a
modal popup allowing the user to set the screenshot resolution (either
in absolute numbers, or scale) and whether the background should be
transparent. Screenshots are saved to the platform specific "pictures"
directory.

The renderer had undergone a serious refactoring to make this
happen. The scene is now always rendered to an intermediate texture
and just then blitted to the backbuffer. This will also allow us to do
postprocessing effects later on. The UI is still drawn directly to the
backbuffer.

Creation of offscreen (as opposed to primary) render
targets (i.e. collection of textures and buffers needed for rendering)
is now possible in the renderer. Each time a screenshot is supposed to
be captured, a new offscreen render target is created, rendered to,
and has the sceenshot picture extracted from it before it is
destroyed. Offscreen render targets are currently multisampled with
the exact same number of samples that the primary render target is
sampled with. This restriction can be lifted later.

Currently not working:

- Vulkan on macOS hangs when trying to render a screenshot (but metal
works)
- Vulkan on Linux segfaults when trying to read the buffer containing
the screenshot if the screenshot has any realistic
dimensions (i.e. greater than 200x200)

* Add notifications window

* Add warning text

* Make notification window smaller