Add configurable shadow rendering
This is just a draft. Don't merge, only evaluate.

The renderer can now draw shadows via shadow mapping. This is used to collect
shadows from rendered meshes and project them onto the ground plane (which is a
plane mesh primitive).

The `CommandBuffer::draw_meshes_*` family of functions now has a flag, whether
the drawing should cast shadows. The `DrawMeshMode` enum has a new
`FlatWithShadows` variant that can collect shadows. The structure is much more
flexible internally in the renderer, but since the renderer does not support
transparency yet, the outside usage is restricted.

Known problems:

- Vulkan validation errors for image layout transitions (this is probably not
  our fault, but it looks scary),
- Screenshots with transparent background can now see ground, because
  it is not transparent. We can alternatively not render the ground plane for
  screenshots,
- The ground is not transparent, objects are not visible from below,
- We needed to disable the ground and shadows for `Edges` shading mode, because
  edges don't write their depth, meaning the ground would overwrite them
- Possibly much more...

Please evaluate this PR and let me know if we want to merge it as-is (or with
minor fixes), not merge it at all, or work on the transparency issues.