Add configurable shadow rendering (#124)
The renderer can now draw shadows via shadow mapping. This is used to collect
shadows from rendered meshes and project them onto the ground plane (which is a
plane mesh primitive).

The CommandBuffer::draw_meshes_* family of functions now has a flag, whether
the drawing should cast shadows. The DrawMeshMode enum has a new
FlatWithShadows variant that can collect shadows. The structure is much more
flexible internally in the renderer, but not yet exposed. Also, the renderer now internally supports transparency, but the only way to render a transparent object is to provide a transparent color to the flat_shading_color option - that color is used for DrawMeshMode::FlatWithshadows.

Known problems:

- Vulkan validation errors for image layout transitions (this is probably not our fault, but it looks scary)