Implement autorun in Session and add it in UI (#137)
This adds a new checkbox to the operations window that controls the interpreter
session "autorun" mode. When enabled the session automatically runs the
interpreter as new edits are made to the program.

Implementation-wise, the `time` variable queried at the beginning of frame
processing is passed down everywhere program edits or polling are made. We could
just use `Instant::now()` internally in those functions, but that does a syscall
at best, and a mutex lock at worst each time (see docs for
`std::time::Instant::now()`). We can potentially simplify later.