Split in-operation analytics, set voxel count limit
Based on real-life use, the software sometimes tries to make heavy calculations. With autorun turned on, it may cause the program to crash. Therefore safety switches have been implemented to prevent computation of heavy geometry.

The analytic operation has been split into bounding box analytics (quick and turned on by default) and mesh analytics (slow and turned off by default).

Voxel based operations are by default checking the resulting voxel count and if it is above the threshold, the operation returns an error, including a suggestion for a new voxel size. The calculation is imprecise.

All these settings are optional and can be toggled.