Adjust texts to match printed manual, add styling to new elements (#150)
Some texts are now returned to their former version, some parts are a compromise between old and new versions (mainly words "sequence" and "pipeline").
The official name of the pipeline is now "Operation pipeline" with a subtitle "sequence of operations".

This is mainly due to a printed manual, that hasn't been updated and still uses the name "Operation pipeline". This compromise can remain in the master.