Display used values as transparent (#147)
The feature is published as a check box in the main menu. Enabled by default,
now all used values/geometries are visible in the viewport, rendered with a
transparent(er) version of the currently selected viewport draw mode.

The renderer is refactored to divorce the concepts of a material and a viewport
draw mode. A viewport draw mode is what the application knows about and maps to
renderer materials. The renderer is now able to render each object within a
batch with a different material. Internally, a material is represented as a set
of values contained in a uniform buffer. Depending on a material's transparency
behavior (opaque/transparent/xray), the renderer automatically selects a correct
render pipeline to draw with - some pipelines read-write depth, some only read
depth, and some always pass the depth test without writing - and sorts the
objects in the three transparency category render lists (opaque, transparent,
xray) based on the object's centroid z value after projection with the camera
view matrix.

The session now notifies the caller with changes to used values. Internal value
diffing algorithm has been simplified, optimized and expanded to support
transitioning values between used and unused value sets.

The "Center viewport" feature now takes into consideration, whether used values
are visible, and always centers only around visible values.