Implement streaming obj serialization
This implements streaming OBJ serialization into a
`std::io::Writer`. Functionality is triggered by a main menu "Export OBJ"
button. Only unused geometries are exported. For each mesh or element of
mesh-array, a new "o" tag is created in the resulting OBJ file. The object name
is the same as the `return_value_name` that would be displayed in the pipeline
window in most cases. For some edge cases however, only the raw stringified var
ident is outputted. If the exported mesh is a mesh-array element and it's index
is not zero, it is part of the name written after the "o" tag.

The commit also does a bit of minor cleanup elsewhere:

- Ensures the session never generates the same variable identifier
  twice (related to being confident that if a var name is present in the
  session, it does not reference a previously removed variable declaration),
- Removes some per-frame uses of the format! macro from `ui.rs`,
- Inlines one simple call to tinyfiledialogs from `ui.rs` to `lib.rs`.