Restore previous numbering of variable names in the pipeline window (#168)
Previously, we used the statement index to generate part of the displayed
variable names in the combo boxes. In the UI however, we did rely on the fact
that the var declaration statement index was the same as the internal raw value
of the variable identifier.

This patch removes that incorrect assumption and once again uses the statement
index to generate the displayed human readable name.