Fix interpreter caching bug (#167)
* Add test that reproduces the bug

This is the most fascinating bug of all, if you know what I mean :)

When we interleave editing the program with executions (what always happens with autorun, but could be achieved manually as well), it is possible to create a variable value in the interpreter that replaces another variable value without the interpreter noticing.

The parameters of the function call that creates that value can potentially reference completely different variable values later on, but since those referenced values were not invalidated, the value of the variable created by referencing them also wasn't.

The fix is to attach a "timestamp" (created_epoch field) to each stored variable value and use that to determine if a variable value could have legally be constructed by referencing another variable value. The other value, while its declaration being earlier in the program, could have been replaced by subsequent program edits and executions.

As a FIXME in the patch suggests, another possible fix could have been to never keep stale values in the interpreter environment and conservatively remove values from the environment whenever the variable declaration statements creating them are edited. This would cause us to loose some caching opportunities, though.

* Patch missing dependency invalidation case