Use clap for command-line argument parsing
In addition to previous configuration via environment variables, we can now
specify command-line arguments.

The reason for this is change is that we sometimes want to configure the
installed application with non-default settings. With just the environment
variables, it is impossible for an application shortcut on Windows to configure
the application (on Windows, env vars cannot be specified inline, but only via
the SET command), but it can set command-line arguments.

I removed the environment variables help section in the readme, because the help
message can how be autogenerated with `-h`:

```
HURBAN selector 0.1.0
Ondrej Slinták <ondrowan@gmail.com>, Ján Tóth <yanchi.toth@gmail.com>, Ján Pernecký <info@sub.digital>

USAGE:
    hurban_selector [OPTIONS]

FLAGS:
    -h, --help       Prints help information
    -V, --version    Prints version information

OPTIONS:
        --gpu-backend <gpu-backend>
            GPU backend to use for rendering [env: HS_GPU_BACKEND=vulkan]  [possible values: vulkan, d3d12, metal]

        --gpu-msaa <gpu-msaa>
            Level of multi-sampling based anti-aliasing to use in rendering [env: HS_GPU_MSAA=]  [default: disabled]
            [possible values: disabled, x2, x4, x8, x16]
        --gpu-power-preference <gpu-power-preference>
            Power preference for selecting a GPU [env: HS_GPU_POWER_PREFERENCE=]  [default: default]  [possible values:
            default, low-power, high-performance]
        --log-level-app <log-level-app>
            Logging level for the editor [env: HS_LOG_LEVEL_APP=]  [default: info]  [possible values: off, error, warn,
            info, debug, trace]
        --log-level-lib <log-level-lib>
            Logging level for external libraries [env: HS_LOG_LEVEL_LIB=]  [default: warn]  [possible values: off,
            error, warn, info, debug, trace]
        --theme <theme>
            Theme for the editor [env: HS_THEME=]  [default: dark]  [possible values: dark, light]
```