Reimplement camera panning in tersms of ray/plane intersections.
Use two intersections of a ray and a plane to determine the vector to move the
camera by instead of using hand-tuned values that only worked for a particular
zoom level range.

The new `Camera::pan_screen` (right mouse button) behaves very well, but we
might want to tweak the place where we put the plane we intersect
with. Currently, that is a plane perpendicular to the screen and passes through
the point the camera is looking at.

The new `Camera::pan_ground` (shift + right mouse button) behaves mostly well,
but if the ray and the plane are too close to being parallel, wild things
happen. For this reason an angle cosine threshold is used in the intersection
computation and if the cosine is too close to zero, the intersection is
discarded. This gives meh experience when using ground panning with sharp
angles, because the camera simply becomes stuck, but at the same time that is
kinda expected.